rename index to size
