+++ src/scenes/Moves/Ppm/PPMSizeWizard.jsx
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import { createOrUpdatePpm, loadPpm } from './ducks';
 import WizardPage from 'shared/WizardPage';
-import PpmSize from '.';
+import PpmSize from './size';
 export class PpmSizeWizardPage extends Component {
   componentDidMount() {
     this.props.loadPpm(this.props.match.params.moveId);
